omg it works

+++ common/Level.zig
@@ -1,7 +1,7 @@
 const Level = @This();
 const math = @import("math.zig");
 
-pub const Tile = struct { pos: math.Point, tile: u16 };
+pub const Tile = struct { pos: math.Point, tile: u8 };
 pub const CompressedMap = struct { levels: []Level, tiles: []Tile };
 pub const CamMode = enum(u8) { locked, follow_x, follow_y, free_follow };
 pub const Entity = struct {
